Added error handling to rsdoslib
I finally sat down and figured out how to swap out the RSDOS
error handler by using a vector to replace it with our own.

You can now check for errors by testing if the Z flag is not set.
If Z is not set after making a call to BASIC or the DOS subroutines
you should check ERRNO. ERRNO is always cleared at the beginning of
a DOS or BASIC call. You can use BASIC_RETURN to put back the original
error handler.